web: wsgi: Do not allow public collections to be removed.
This commit implements checks to prevent removing public collections.

* src/djehuty/web/resources/html_templates/depositor/my-collections.html:
  Don't show the delete button for public collections.
* src/djehuty/web/resources/sparql_templates/delete_collection.sparql:
  Protect deleting public collections at the query level.
* src/djehuty/web/wsgi.py: Protect deleting public collections at the
  API level.

+++ src/djehuty/web/wsgi.py
@@ -830,6 +830,9 @@ class ApiServer:
                         account_id    = account_id)[0]
 
                     collection_version_id = collection["collection_version_id"]
+                    if collection["is_public"] != 0:
+                        return self.error_403 (request)
+
                     result = self.db.delete_collection (
                         collection_version_id = collection_version_id,
                         account_id    = account_id)
